Initial implementation of config
The basic approach is to add WEB-INF/ibm-web-bnd.xml file
to the application to bind it to a specific virtual host.

The binding is based off the hash of the application location.
This does have a limitation that an application can only be
loaded, at most, once for a specific location.  A separate
bundle is then installed to represent the configuration of
the application by using a OSGI-INF/wlp/defaultInstances.xml
file that contains the application specific configuration
for the virtual host.