When allowing multiple versions pull in all tolerated versions
The allowMultipleVersions flag is used by minify to allow multiple
versions of a feature to be included in a packaged server.  Before this
change only the preferred versions of included features were pulled into
the packaged server.  This could result in a packaged server that does
not include all the necessary features to start the server.

This change also pulls in all tolerated candidate features so that we
can gaurantee at least the minimum set of features required to start the
configured server in included.