Ensure messages about application starting happen on restore
When restoring a checkpoint server that has already started
applications the application starting messages will not
be posted for each server restore run.

This could be confusing to users not to see messages
indicating their application is started.  It also poses
issues for our FAT tests that verify the configured
applications are started.

To address this the ApplicationConfigurator checkpoint
hook is enhanced to collect application messages that
happen pre-checkpoint and post them on resore.