Implement CRaC APIs on Liberty checkpoint
Initial, no-ship, crac-1.0 feature that implements
org.crac third-party APIs that allows applications
to participate in the checkpoint process.

Add APP_REQUESTED checkpoint phase. If APP_REQUESTED
phase is used and no app request is made to checkpoint
then checkpoint will fail. If an application requested
checkpoint is made but the APP_REQUESTED checkpoint
phase is not used then an exception is thrown back
to the caller and no checkpoint is taken.